Deprecate #use_vcr_cassette
- This commit fixes #212
